chore: change uint fields to int
The PR #8662 added validation markers for all unsigned integer values.
We can now safely change the Go types to comply with the Kubernetes API
conventions which strongly recommend signed integers in Go definitions.

+++ pkg/prometheus/validation/validator_test.go
@@ -372,7 +372,7 @@ func TestValidateRelabelConfig(t *testing.T) {
 				TargetLabel:  "__port1",
 				Separator:    new(relabel.DefaultRelabelConfig.Separator),
 				Regex:        relabel.DefaultRelabelConfig.Regex.String(),
-				Modulus:      relabel.DefaultRelabelConfig.Modulus,
+				Modulus:      int64(relabel.DefaultRelabelConfig.Modulus),
 				Replacement:  &relabel.DefaultRelabelConfig.Replacement,
 				Action:       "keepequal",
 			},
